docs/scripts/_ia/requirements.txt
# scripts/_ia/ Dependencies
# Python packages for AI infrastructure: extractors, validators, generators, agents, utils

# Code quality and complexity analysis
radon>=6.0.1          # Cyclomatic complexity, LOC metrics
lizard>=1.17.10       # Code complexity and metrics (alternative to radon)

# YAML and JSON processing
PyYAML>=6.0.1         # YAML parsing and generation
jsonschema>=4.21.0    # JSON schema validation

# CLI utilities
click>=8.1.7          # Command-line interface creation

# Optional: type checking (for strict mode)
mypy>=1.8.0           # Static type checker

# Optional: linting
pylint>=3.0.0         # Code linting
autopep8>=2.0.0       # Code formatting (SSOT-safe)
